Crate copperline

A low-level terminal line editing library

Copperline is a line editing library written from scratch in Rust, born from the authors frustration with the interface of existing bindings to C-libraries like readline, libedit or linenoise.

Features that are currently supported:

Cursor movement and text insertion

Jumps (via C-a and C-e)

History browsing (via C-p and C-n)

It has a clean, hackable codebase, which I hope will foster contributions so that the Rust ecosystem will soon be able to utilise a mature, pure-Rust line editing library.

Structs

Copperline

Enums

Encoding

Error